Debug the tooltip enable / disable problem + add prefabs for TooltipManager + LocalizManager + ClassEvolutionManager + CharacterManager

+++ Assets/Scripts/Characters/CharacterStats.cs
@@ -3,13 +3,13 @@ using UnityEngine;
 
 public class CharacterStats
 {
-    public Dictionary<StatType, int> stats = new();
+    public Dictionary<StatType, float> stats = new();
 
     public CharacterStats(StatProfile profile)
     {
         foreach (var s in profile.stats)
             stats[s.type] = s.proficiency;
     }
-    public int GetProficiency(StatType type) => stats.ContainsKey(type) ? stats[type] : 0;
+    public float GetProficiency(StatType type) => stats.ContainsKey(type) ? stats[type] : 0;
 
 }
